requirements.txt
numpy
pandas
scikit-learn
matplotlib
spacy
stanza
faster-whisper
openai-whisper
pyyaml
tqdm
